reworked documentation restructured authentication layer
