+++ .github/workflows/ci.yml
@@ -108,27 +108,27 @@ jobs:
     - name: Build docs ⚒️
       run: make docs
 
-  deploy:
-    name: Deploy to PyPI
-    runs-on: ubuntu-latest
+  # deploy:
+  #   name: Deploy to PyPI
+  #   runs-on: ubuntu-latest
 
-    # deploy will only be run when there is a tag available
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-    needs: [quality, test, docs]  # only run after all other stages succeeded
+  #   # deploy will only be run when there is a tag available
+  #   if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+  #   needs: [quality, test, docs]  # only run after all other stages succeeded
 
-    steps:
-    - name: Checkout git repository 🕝
-      uses: actions/checkout@v2
+  #   steps:
+  #   - name: Checkout git repository 🕝
+  #     uses: actions/checkout@v2
 
-    - name: Set up python3.8 🐍
-      uses: actions/setup-python@v2
-      with:
-        python-version: 3.8
+  #   - name: Set up python3.8 🐍
+  #     uses: actions/setup-python@v2
+  #     with:
+  #       python-version: 3.8
 
-    - name: Install poetry 🦄
-      uses: Gr1N/setup-poetry@v4
+  #   - name: Install poetry 🦄
+  #     uses: Gr1N/setup-poetry@v4
 
-    - name: Build ⚒️ distributions
-      run: |
-        poetry build
-        poetry publish -u __token__ -p ${{ secrets.PYPI_TOKEN }}
+  #   - name: Build ⚒️ distributions
+  #     run: |
+  #       poetry build
+  #       poetry publish -u __token__ -p ${{ secrets.PYPI_TOKEN }}
