+++ capella_console_client/__init__.py
@@ -1,1 +1,1 @@
-from .client import CapellaConsoleClient  # flake8: noqa
+from .client import CapellaConsoleClient
