+++ docs/pages/installation.rst
@@ -7,7 +7,7 @@ Installation
 Use a Published Release
 #######################
 
-To install `capella-console-client`, simply run this command in your terminal of
+To install ``capella-console-client``, simply run this command in your terminal of
 choice:
 
 .. code-block:: console
@@ -17,10 +17,6 @@ choice:
 Build from Source
 #################
 
-`developed on GitHub <https://github.com/capellaspace/console-client>`_.
-
-You can either clone the public repository:
-
 .. code-block:: console
 
   $ git clone git@github.com:capellaspace/console-client.git
@@ -29,17 +25,18 @@ Or, via tarball:
 
 .. code-block:: console
 
-  $ curl -OL https://github.com/capellaspace/console-client/tarball/master
+  $ curl -OL https://github.com/capellaspace/console-client/tarball/main
 
 .. note::
   If you are using windows, you can also download a zip instead:
 
   .. code-block:: console
 
-    $ curl -OL https://github.com/capellaspace/console-client/zipball/master
+    $ curl -OL https://github.com/capellaspace/console-client/zipball/main
 
-`capella_console_client` uses `Poetry <https://python-poetry.org/>`_ for packaging and
-dependency management. If you want to build `capella_console_client` from source, you
+
+``capella-console-client`` uses `Poetry <https://python-poetry.org/>`_ for packaging and
+dependency management. If you want to build ``capella-console-client`` from source, you
 need to install Poetry first:
 
 .. code-block:: console
@@ -49,7 +46,7 @@ need to install Poetry first:
 There are several other ways to install Poetry, as seen in
 `the official guide <https://python-poetry.org/docs/#installation>`_.
 
-To install capella_console_client and its dependencies in editable mode, execute
+To install ``capella-console-client`` and its dependencies in editable mode, execute
 
 .. code-block:: console
