+++ docs/pages/support.rst
@@ -4,5 +4,5 @@ Support
 
 Please `open an issue <https://github.com/capellaspace/console-client/issues/new>`_
 with enough information for us to reproduce your problem.
-A `minimal, reproducible example <https://stackoverflow.com/help/minimal-reproducible-example>`_
-would be very helpful.
+
+Please provide a `minimal, reproducible example <https://stackoverflow.com/help/minimal-reproducible-example>`_.
